+++ packages/fluentui/react-northstar/src/themes/teams/components/Menu/submenuIndicatorDirection.ts
